add develop branch and fix function
