+++ controller/citizen.js
@@ -37,42 +37,25 @@ exports.citizen_mgr = {
   },
   add_citizen: async body => {
     try {
-      let citizen = await models.Citizen.findOne({
-        where: {
-          person_id: body.person_id
-        }
-      });
-      if (citizen) {
-        let tempcitizen = {
-          person_id: citizen.person_id,
-          national_id: citizen.national_id,
-          first_name: citizen.first_name,
-          father_name: citizen.father_name,
-          grand_name: citizen.grand_name,
-          family_name: citizen.family_name,
-          mother_name: citizen.mother_name,
-          birth_date: citizen.birth_date,
-          gender: citizen.gender,
-          is_alive: citizen.is_alive,
-          registry_number: citizen.registry_number,
-          deleted: citizen.deleted
-
-        }
-        await models.tempCitizen.create(tempcitizen);
-        await models.Citizen.destroy({
-          where: {
-            person_id: body.person_id
-          }
-        });
-        return await models.Citizen.create(tempcitizen);
-      } else {
-        return await models.Citizen.create(tempcitizen);
-      }
+      return await models.Citizen.create(body);
 
     } catch (err) {
       return (err);
     }
 
   },
+  delete_citizen: async body => {
+    try {
+      return await models.Citizen.destroy({
+        where: {
+          person_id: body.person_id
+        }
+      });
 
+    } catch (err) {
+      return (err);
+    }
+
+
+  },
 };
